+++ tests/integration/conftest.py
@@ -78,6 +78,16 @@ def int_table2(engine, metadata):
 
 
 @pytest.fixture(scope="module")
+def int_square_table(engine, metadata):
+    table_name = "int_square_table"
+    engine.execute(f"DROP TABLE IF EXISTS {SCHEMA}.{table_name}")
+    columns = [sa.Column("col_int", sa.Integer())]
+    data = [{"col_int": i**2} for i in range(1, 20)]
+    _handle_table(engine, metadata, table_name, columns, data)
+    return TEST_DB_NAME, SCHEMA, table_name
+
+
+@pytest.fixture(scope="module")
 def mix_table1(engine, metadata):
     table_name = "mix_table1"
     columns = [