+++ tests/integration/test_integration.py
@@ -1755,3 +1755,54 @@ def test_groupby_aggregation_within_with_failures(
 
 def test_diff_average_between():
     return
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        (identity, "col_int", "col_int", None, 1.0),
+        (identity, "col_int", "col_int", Condition("col_int >= 3"), 1.0),
+    ],
+)
+def test_ks_2sample_constraint_perfect_between(engine, int_table1, data):
+    """
+    Test Kolmogorov-Smirnov for the same column -> p-value should be perfect 1.0.
+    """
+    (operation, col_1, col_2, condition, significance_level) = data
+    req = requirements.BetweenRequirement.from_tables(*int_table1, *int_table1)
+    req.add_ks_2sample_constraint(
+        column1=col_1,
+        column2=col_2,
+        condition1=condition,
+        condition2=condition,
+        significance_level=significance_level,
+    )
+
+    assert operation(req[0].test(engine).outcome)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        (negation, "col_int", "col_int", 0.05),
+        (
+            identity,
+            "col_int",
+            "col_int",
+            0.0,
+        ),  # test should succeed although data is different
+    ],
+)
+def test_ks_2sample_constraint_wrong_between(
+    engine, int_table1, int_square_table, data
+):
+    """
+    Test kolmogorov smirnov test for table and square of table -> significance level should be less than default 0.05
+    """
+    (operation, col_1, col_2, min_p_value) = data
+    req = requirements.BetweenRequirement.from_tables(*int_table1, *int_square_table)
+    req.add_ks_2sample_constraint(
+        column1=col_1, column2=col_2, significance_level=min_p_value
+    )
+
+    assert operation(req[0].test(engine).outcome)