+++ CHANGELOG.rst
@@ -7,6 +7,13 @@
 Changelog
 =========
 
+1.1.0 - 2022.06.01
+------------------
+
+**New feature: Statistical Tests**
+
+- Implemented a new constraint :class:`~datajudge.constraints.stats.KolmogorovSmirnov2Sample` for :class:`~datajudge.BetweenRequirement` that performs a `Kolmogorov Smirnov Test <https://en.wikipedia.org/wiki/Kolmogorov%E2%80%93Smirnov_test>`_ between two data sources.
+
 1.0.1 - 2022.05.24
 ------------------